refactor(QoL): :recycle: componentize Sidebar by splitting Header, Footer, and Content
